Refactor: extract programme detail modal into web/js/components/modal.js (#77)
Moves openProgrammeModal, openSearchAiringModal, closeModal, and their
private helpers (setVisible, addBadge) out of main.js into a dedicated
component module. Updates sw.js to pre-cache the new file.
